refactor(admin): extract WorkflowsTab into its own file
